# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Unlock Content Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List and search leads
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead with full history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDetail"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadDetail"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/{id}/next-action:
    get:
      operationId: getLeadNextAction
      tags: [leads]
      summary: Get next best action for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Next best action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NextAction"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Lead:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        company:
          type: ["string", "null"]
        pipeline_stage:
          type: string
        first_contact:
          type: string
        last_contact:
          type: string
        detected_persona:
          type: ["string", "null"]
        archived:
          type: boolean
        send_count:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - name
        - pipeline_stage
        - first_contact
        - last_contact
        - archived
        - send_count
        - created_at
        - updated_at
    LeadDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        company:
          type: ["string", "null"]
        pipeline_stage:
          type: string
        first_contact:
          type: string
        last_contact:
          type: string
        detected_persona:
          type: ["string", "null"]
        archived:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
        send_log:
          type: array
          items:
            $ref: "#/components/schemas/SendLogEntry"
        stage_history:
          type: array
          items:
            $ref: "#/components/schemas/StageHistoryEntry"
        notes:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - pipeline_stage
        - first_contact
        - last_contact
        - archived
        - created_at
        - updated_at
        - send_log
        - stage_history
        - notes
    CreateLeadBody:
      type: object
      properties:
        name:
          type: string
        company:
          type: ["string", "null"]
        pipeline_stage:
          type: string
          default: "Outreach"
      required:
        - name
    UpdateLeadBody:
      type: object
      properties:
        pipeline_stage:
          type: string
        detected_persona:
          type: ["string", "null"]
        archived:
          type: boolean
        company:
          type: ["string", "null"]
    SendLogEntry:
      type: object
      properties:
        send_id:
          type: string
        date:
          type: string
        documents_sent:
          type: array
          items:
            type: string
        pipeline_stage_at_send:
          type: string
        transcript_summary:
          type: string
        analysis_confidence:
          type: object
          properties:
            persona:
              type: number
            stage:
              type: number
          required:
            - persona
            - stage
        already_sent_exclusions:
          type: array
          items:
            type: string
        blocked_documents_seen:
          type: array
          items:
            type: string
        stage_suggestion:
          type: ["string", "null"]
        email_sent:
          type: boolean
        status:
          type: string
      required:
        - send_id
        - date
        - documents_sent
        - pipeline_stage_at_send
        - transcript_summary
        - analysis_confidence
        - already_sent_exclusions
        - blocked_documents_seen
        - email_sent
        - status
    StageHistoryEntry:
      type: object
      properties:
        stage:
          type: string
        date:
          type: string
        logged_by:
          type: string
      required:
        - stage
        - date
        - logged_by
    NextAction:
      type: object
      properties:
        action:
          type: string
        rationale:
          type: string
        suggested_documents:
          type: array
          items:
            type: string
      required:
        - action
        - rationale
        - suggested_documents
...