# lib/api-zod/src/generated/types/leadDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unlock Content Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { SendLogEntry } from "./sendLogEntry";
import type { StageHistoryEntry } from "./stageHistoryEntry";

export interface LeadDetail {
  id: string;
  name: string;
  /** @nullable */
  company?: string | null;
  pipeline_stage: string;
  first_contact: string;
  last_contact: string;
  /** @nullable */
  detected_persona?: string | null;
  archived: boolean;
  created_at: string;
  updated_at: string;
  send_log: SendLogEntry[];
  stage_history: StageHistoryEntry[];
  notes: string[];
}