# lib/api-zod/src/generated/types/updateLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unlock Content Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateLeadBody {
  pipeline_stage?: string;
  /** @nullable */
  detected_persona?: string | null;
  archived?: boolean;
  /** @nullable */
  company?: string | null;
}